Adds documentation for undocumented `withSlide`
I had a use-case where I needed access to the "active" state of the current slide so I could reset the animation in a custom component. After digging around in the source I discovered the undocumented `withDeck` and `withSlide` functions. I couldn't distinguish any difference between the two when I tried them, but the `withSlide` function was exactly what I needed.

This pull request just adds a line to `advanced.md` describing how to use `withSlide` to get access to the current active state of the slide.