Added detailed example of automatic transitions
Hi, great library... thanks! I needed to create a slide deck which advances automatically and though it might be helpful to share with others. The one issue this implementation has is that the timers do not reset if the user manually advances the presentation, causing some flickering. Thanks!

You can see an example of this in use, here: [https://color-test-mwg705juq.now.sh](https://color-test-mwg705juq.now.sh)